new order stored procedure updated

+++ stored-procedure.sql
@@ -57,6 +57,7 @@ IF OBJECT_ID('NewOrder', 'P') IS NOT NULL
 DROP PROCEDURE NewOrder; 
 GO 
 
+
 CREATE PROCEDURE NewOrder
 	@customerID int,	
 	@oneShipment BIT, 
@@ -70,50 +71,56 @@ DECLARE @currentDate varchar(10)
 DECLARE @shippingdate varchar(10)
 DECLARE @productArray table(product int) 
 DECLARE @quantityArray table (quantity int)
+DECLARE @currentBalance decimal(10,2)
+DECLARE @creditLimit decimal(10,2)
+DECLARE @defaultCreditCard varchar(19)
+
+SET @defaultcreditCard = (select DefaultCreditCard from Customer where CustomerID = @customerID) 
+SET @currentBalance = (SELECT CurrentBalance from Balance where CustomerID = @customerID)
+SET @creditLimit = (SELECT CreditLimit from Balance where CustomerID = @customerID)
 SET @currentDate = ( SELECT CONVERT (date, SYSDATETIME()) )
 SET @shippingdate = ( SELECT DATEADD(day, 5, (SELECT CONVERT (date, SYSDATETIME()))))
+If (@defaultcreditCard is NULL) AND (@orderTotalPrice + @currentBalance > @creditLimit)
+	BEGIN
+		RETURN 0
+	END
+ELSE --CREDIT CARD EXISTS OR IF CREDIT CARD DOES NOT EXIST BUT order is within customer's credit limits
+	begin
 
-Insert into Invoice( CustomerID, OneShipment, OrderTotalPrice, EstimatedShippingDate, InvoiceDateTime)
-Values (@customerID, @oneShipment, @orderTotalPrice, @shippingdate, @currentDate)
+		Insert into Invoice( CustomerID, OneShipment, OrderTotalPrice, EstimatedShippingDate, InvoiceDateTime)
+		Values (@customerID, @oneShipment, @orderTotalPrice, @shippingdate, @currentDate)
 
-INSERT INTO @productArray
-SELECT CAST(value AS int) FROM STRING_SPLIT(@products, ',');
+		INSERT INTO @productArray
+		SELECT CAST(value AS int) FROM STRING_SPLIT(@products, ',');
 
-INSERT INTO @quantityArray
-SELECT CAST (value AS int) FROM STRING_SPLIT(@quantities, ',');
+		INSERT INTO @quantityArray
+		SELECT CAST (value AS int) FROM STRING_SPLIT(@quantities, ',');
 
-DECLARE @count int = 1;
+		DECLARE @count int = 1;
 
-DECLARE @length int;
-SELECT @length = COUNT(*) FROM @productArray
+		DECLARE @length int;
+		SELECT @length = COUNT(*) FROM @productArray
 
-DECLARE @invoice int;
-SELECT @invoice = InvoiceID FROM Invoice WHERE CustomerID = @customerID
-WHILE @count <= @length
-BEGIN
-	DECLARE @product int;
-	DECLARE @quantity int;
+		DECLARE @invoice int;
+		SELECT @invoice = InvoiceID FROM Invoice WHERE CustomerID = @customerID
+		WHILE @count <= @length
+		BEGIN
+			DECLARE @product int;
+			DECLARE @quantity int;
 
-	SELECT TOP 1 @product = product FROM @productArray;
-	SELECT TOP 1 @quantity = quantity FROM @quantityArray;
+			SELECT TOP 1 @product = product FROM @productArray;
+			SELECT TOP 1 @quantity = quantity FROM @quantityArray;
 
-	INSERT INTO InvoiceLineItems (InvoiceID, InvoiceLineNO, InvoiceLineItemAmount, ProductID, Quantity)
-	SELECT @invoice, @count, BestPrice, @product, @quantity FROM Inventory WHERE InventoryID = @product;
+			INSERT INTO InvoiceLineItems (InvoiceID, InvoiceLineNO, InvoiceLineItemAmount, ProductID, Quantity)
+			SELECT @invoice, @count, BestPrice, @product, @quantity FROM Inventory WHERE InventoryID = @product;
 
-	DELETE FROM @productArray WHERE product = @product;
-	DELETE FROM @quantityArray WHERE quantity = @quantity;
+			DELETE FROM @productArray WHERE product = @product;
+			DELETE FROM @quantityArray WHERE quantity = @quantity;
 
-	SET @count = @count + 1;
-END
+			SET @count = @count + 1;
+		END
 
-END
+		END
+	END
+
 GO
-
---select * from Invoice;
-
---testing exec
---EXEC dbo.NewOrder
---@customerID = 1, @oneShipment = 0, @orderTotalPrice = 12.56, @products = '1,2,3', @quantities = '2,1,5'
---GO
---SELECT * FROM Invoice;
---SELECT * FROM InvoiceLineItems;
